WPI: Added cmd_vel subscriber. Builds but doesn't work

+++ rover/raspi_rover/raspi_robot_driver/src/raspi_robot_servo.c
@@ -43,10 +43,10 @@ int16_t servo_set(uint8_t gpio_num, int16_t angle) {
   // Angle in range min to max maps on to duty 0 to 255.
   // Convert +90 to -90 angle to 0 to 180.
   uint16_t single_angle = angle + k_angle_offset;
-  uint8_t duty = (uint8_t)(single_angle * 256) / 180;
+  uint16_t duty = (single_angle * 256) / 180;
     // Set duty.
   led_pwm_handle_p handle = led_pwm_get_handle(gpio_num);
-  led_pwm_set_duty(handle, duty);
+  led_pwm_set_duty(handle, (uint8_t)duty);
   ESP_LOGI(TAG, "Servo set GPIO %d to %d degrees using %d duty.", gpio_num, angle, duty);
   return angle;
 }
